Fixes problem with symlinked modules (#81)
If you npm link a dependency, and it has dependencies with asset-manifest.json files, they will not be found because glob won't traverse the symlinked directory. This was introduced in #78.
This fixes that problem, but slows down startup times. One of our larger projects startup time goes from 8 seconds to 20 seconds with this change.